api/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
docling_core
Pillow
aiofiles